move handlePreload into its own file
