+++ src/hooks/server/handlePreload.ts
@@ -1,0 +1,13 @@
+import type { Handle } from "@sveltejs/kit";
+
+const handlePreload = (async ({ event, resolve }) => {
+  const response = await resolve(event, {
+    preload: ({ type, path }) =>
+      type === "js" ||
+      type === "css" ||
+      (type === "font" && !!path.match(/sourcesanspro-regular-webfont.[0-9a-z]{8}.woff2$/)),
+  });
+  return response;
+}) satisfies Handle;
+
+export default handlePreload;
